% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell.proportions}
\alias{cell.proportions}
\title{Cell Proportion data
Example cell proportions of circulating immune cell subpopulations from 5 healthy individuals}
\format{A matrix with 33 quantified cell proportions 5 samples}
\usage{
cell.proportions
}
\description{
Cell Proportion data
Example cell proportions of circulating immune cell subpopulations from 5 healthy individuals
}
\keyword{datasets}
