% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count.table}
\alias{count.table}
\title{Expression data
Example RNA-seq data from 5 healthy individuals}
\format{A matrix with 42992 quantified genes and 5 samples}
\usage{
count.table
}
\description{
Expression data
Example RNA-seq data from 5 healthy individuals
}
\keyword{datasets}
