% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.evaluate}
\alias{dCell.evaluate}
\title{Evaluate percentage of coeficient used in a model
\code{dCell.predict} Predicts cell proportions of 32 cell subpopulations usin decon Cell models.}
\usage{
dCell.evaluate(genes.ensembl, ct.model)
}
\arguments{
\item{genes.ensembl}{Character vector of ensemble gene ids overlapping between the decon cell model and the new gene expression.}

\item{ct.model}{A single decon cell model. A numeric vector with ensembl IDs as names.}
}
\value{
Percentage of coeficient represented in the new data
}
\description{
Evaluate percentage of coeficient used in a model
\code{dCell.predict} Predicts cell proportions of 32 cell subpopulations usin decon Cell models.
}
\examples{
dCell.evaluate(genes.ensembl, ct.model)
}
