% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.exp.trimming}
\alias{dCell.exp.trimming}
\title{Trims a dCell.exp matrix for only dCell genes
\code{dCell.predict} Loads a the gene list used in all dCell models and trims a dCell.exp matrix.}
\usage{
dCell.exp.trimming(dCell.exp)
}
\arguments{
\item{dCell.exp}{A matrix of preprocessed gene expression data}
}
\value{
Percentage of coeficient represented in the new data
}
\description{
Trims a dCell.exp matrix for only dCell genes
\code{dCell.predict} Loads a the gene list used in all dCell models and trims a dCell.exp matrix.
}
\examples{
dCell.exp.trimming(dCell.exp)
}
