% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dCell.geneList}
\alias{dCell.geneList}
\title{DeconCell gene list
Vector of gene names (ENSEMBL identifiers) used in all of the Decon cell models}
\format{Lists of 100 elemented with coefficients, names are coded in ENSEMBL identifiers}
\usage{
dCell.geneList
}
\description{
DeconCell gene list
Vector of gene names (ENSEMBL identifiers) used in all of the Decon cell models
}
\keyword{datasets}
