% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dCell.models}
\alias{dCell.models}
\title{DeconCell models
Lists fo 33 lists containing the coefficients used in the prediction of cell proportions using expression values}
\format{Lists of 100 elemented with coefficients, names are coded in ENSEMBL identifiers}
\usage{
dCell.models
}
\description{
DeconCell models
Lists fo 33 lists containing the coefficients used in the prediction of cell proportions using expression values
}
\keyword{datasets}
