% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dCell.names}
\alias{dCell.names}
\title{Cell subpopulation names
Code and names for the predicted cell subpopulations in the DeconCell package}
\format{data.frame of 33 rows and 2 columns, rownames are the codes foe the dCell.models lists}
\usage{
dCell.names
}
\description{
Cell subpopulation names
Code and names for the predicted cell subpopulations in the DeconCell package
}
\keyword{datasets}
