% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.predict}
\alias{dCell.predict}
\title{Predict all available cell subpopoulation in DeconCell
\code{dCell.predict} Runs dCell.predict}
\usage{
dCell.predict(dCell.exp, dCell.models, res.type = "median",
  custom = FALSE)
}
\arguments{
\item{dCell.exp}{A matrix or data.frame of normalized (by gene lenght and library size (RPKM, TMM, TPM...)) and scaled expression data, samples by columns and genes by row. Preferably comming from \code{dCell.expProcessing}.}

\item{dCell.models}{A list containing the DeconCell models.}

\item{res.type}{Whether the prediction per DeconCell models should be sumarized by either \code{median} or \code{mean}.}
}
\value{
A matrix of predicted cell proportions for all available cell subpopoulation in DeconCell.

A vector containing the performance evaluation of all available models.
}
\description{
Predict all available cell subpopoulation in DeconCell
\code{dCell.predict} Runs dCell.predict
}
\examples{
data(dCell.models)
x <- dCell.predict(dCell.exp, dCell.models=dCell.models)
x$dCell.prediction
x$Evaluation
}
