% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.predict.single}
\alias{dCell.predict.single}
\title{Predict a single cell subpopoulation using a DeconCell model
\code{dCell.predict.single} Predicts cell proportions of one out of the 32 cell subpopulations using the DeconCell models.}
\usage{
dCell.predict.single(dCell.exp, dCell.model, res.type = "median")
}
\arguments{
\item{dCell.exp}{A table of normalized and scaled expression, samples by columns and genes by row. Preferably comming from \code{dCell.expProcessing}.}

\item{res.type}{Whether the prediction per DeconCell models should be sumarized by either \code{median} or \code{mean}.}

\item{dCell.models}{A list containing the DeconCell models.}
}
\value{
A vector of predicted cell proportions for a given set of models.

Mean of percentage of coeficients covered given
}
\description{
Predict a single cell subpopoulation using a DeconCell model
\code{dCell.predict.single} Predicts cell proportions of one out of the 32 cell subpopulations using the DeconCell models.
}
\examples{
data(dCell.models)
x <- dCell.predict.single(dCell.exp, dCell.model=dCell.models[[1]])
x$Predicted
x$Prediction.eval
}
