% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.run}
\alias{dCell.run}
\title{Run Decon-cell to generate new gene signature and characterize marker genes.
\code{dCell.run} Runs dCell.run to generate gene expression signatures to predict cell proportions of bulk tissues}
\usage{
dCell.run(exp, proportions, iterations = 10)
}
\arguments{
\item{exp}{A matrix or data.frame of normalized (by gene lenght and library size (RPKM, TMM, TPM...)) and scaled expression data, samples by columns and genes by row. Preferably comming from \code{dCell.expProcessing}.}

\item{proportions}{A matrix or data.frame of cell proportions within each bulk RNA-Seq sample. Samples by row and cell types by column.}

\item{iterations}{Number of iterations (and models) which would be calculated}

\item{test.size}{Proportion of samples to be considered as test, 100 - \code{test.size} would be considered de proportion used for training.}
}
\value{
a dCell.models-like object, a list (with a \code{length == ncol(proportion.mat))}, each element of the list contains the number of models defined by \code{iterations}
}
\description{
Run Decon-cell to generate new gene signature and characterize marker genes.
\code{dCell.run} Runs dCell.run to generate gene expression signatures to predict cell proportions of bulk tissues
}
\examples{
dCell.run()
}
