% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.run.single}
\alias{dCell.run.single}
\title{Run DeconCell to generate new gene signature for a single cell type.
\code{dCell.run.single} Runs dCell.run to generate gene expression signatures to predict cell proportions of bulk tissues}
\usage{
dCell.run.single(exp, i.ct.proportions, n.samples.train,
  n.samples.test = NULL, iterations = 10)
}
\arguments{
\item{exp}{A matrix or data.frame of normalized (by gene lenght and library size (RPKM, TMM, TPM...)) and scaled expression data, samples by columns and genes by row. Preferably comming from \code{dCell.expProcessing}.}

\item{i.ct.proportions}{A matrix or data.frame of cell proportions within each bulk RNA-Seq sample. Samples by row and cell types by column.}
}
\description{
Run DeconCell to generate new gene signature for a single cell type.
\code{dCell.run.single} Runs dCell.run to generate gene expression signatures to predict cell proportions of bulk tissues
}
