% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{get.marker.genes}
\alias{get.marker.genes}
\title{Wrapper function for glmnet
\code{get.marker.genes}}
\usage{
get.marker.genes(marker.gene.list, iterations,
  frequency.percentage = 0.7)
}
\arguments{
\item{coef.list}{a list of character vectors}
}
\value{
A character vector with all marker genes for a dCell.run
}
\description{
Wrapper function for glmnet
\code{get.marker.genes}
}
